FunctionProjTest - Copy/requirements.txt
pyimport==0.5.6
azure-functions
astroid==2.3.3
azure-functions-worker==1.1.0
azure-storage-file-datalake==12.0.1
azure-core==1.4.0
azure-identity==1.3.1
azure-storage-blob==12.3.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
grpcio==1.26.0
grpcio-tools==1.26.0
idna==2.9
isort==4.3.21
isodate==0.6.0
oauthlib==3.1.0
requests-oauthlib==1.3.0
msrest==0.6.13
pycparser==2.20
cffi==1.14.0
cryptography==2.9.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.15.4
protobuf==3.6.1
pylint==2.4.4
requests==2.23.0
six==1.14.0
urllib3==1.25.9
wrapt==1.11.2
# YAML
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
PyYAML==5.3.1
coloredlogs==14.0
# SPLUNK
splunk-sdk==1.6.12
# micropython-xml.dom==0.0.0
httplib2==0.18.1